I removed the contact form from my website and replaced it with icons for a more minimal and better design. This makes the website less cluttered and easier to navigate.
I changed the color of the menu-bar button to improve the overall visual appeal of the website. The new color is more in line with the website's color scheme and gives a better impression to the user.

I added an href=#contact to the "Hire Me" button, so that users can quickly jump to the contact section of the website. This makes it easier for potential clients to get in touch with me and increases the chances of getting hired.